Hien

+++ server/src/services/user.js
@@ -300,32 +300,23 @@ export const unBanUser = ({ id }) => {
   });
 };
 
-export const addWishlist = ({ id, list }) => {
+export const addWishlist = ({ id, projectID }) => {
   return new Promise(async (resolve, reject) => {
     try {
-      let listdb = list.split(',')
-      let array = []
-      // listdb.forEach((item) => {
-      //   array.push({userID:id, projectID:item})
-      // });
-      for (let i = 0; i < listdb.length; i++) {
-        const wishList = await db.WishList.findOne({
-          where: {
-            userID: id,
-            projectID: listdb[i]
-          }
-        })
-        if (!wishList) {
-          array.push({ userID: id, projectID: listdb[i] })
+      const [wishlist, created] = await db.WishList.findOrCreate({
+        where: {
+          userID: id,
+          projectID: projectID,
+        },
+        defaults: {
+          userID: id,
+          projectID: projectID
         }
-      }
-      const wishlist = await db.WishList.bulkCreate(array, {
-        returning: true
       })
       resolve({
-        err: wishlist ? 0 : 1,
-        mess: wishlist ? "Add to wishlist success" : "Can not add project to wishlist",
-        data: wishlist ? wishlist : ""
+        err: created ? 0 : 1,
+        mess: created ? "Add to wishlist success." : "You have already added this project to wishlist!",
+        data: created ? wishlist : ""
       })
     } catch (err) {
       console.log(err);
@@ -412,30 +403,25 @@ export const viewwishlist = ({ id, page, limit, orderBy, orderType }) => {
   });
 };
 
-export const deletewishlist = ({ id, list }) => {
+export const deletewishlist = ({ id, projectID }) => {
   return new Promise(async (resolve, reject) => {
     try {
-      let listdb = list.split(',')
-      let array = []
-      for (let i = 0; i < listdb.length; i++) {
-        const wishlist = await db.WishList.findOne({
-          where: {
-            userID: id,
-            projectID: listdb[i]
-          }
-        })
-        if(wishlist){
-        array.push(wishlist.id)
-        }
-      }
-      await db.WishList.destroy({
+      const wishList = await db.WishList.findOne({
         where: {
-          id: array
+          userID: id,
+          projectID,
         }
       })
+      if (wishList) {
+        await db.WishList.destroy({
+          where: {
+            id: projectID
+          }
+        })
+      }
       resolve({
-        err: array ? 0 : 1,
-        mess: array ? "Delete project from wishlist success" : "Can not delete project from wishlist"
+        err: wishList ? 0 : 1,
+        mess: !wishList ? `Project (${projectID}) does not belong to User (${id})!` : "Delete project from wishlist successfully."
       })
     } catch (err) {
       console.log(err);
